+++ mystring.h
@@ -25,7 +25,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 void split_get_params(connection_t* connection_prop);
 bool endsWith(char *str, char *end,size_t len_str,size_t len_end);
-void delChar(char *string, int pos, int n);
+void delChar(char *string, int pos, size_t n);
 void strReplace(char *string, char *substr, char with);
 void replaceEscape(char *string);
 void strToUpper(char *str);
